Move walkers related egui windows to map_windows.rs
