+++ crates/re_space_view_map/src/map_space_view.rs
@@ -1,9 +1,8 @@
 use egui::{Color32, TextEdit};
-use egui::{Ui, Window};
 
 use re_types::components::{Color, Radius};
 
-use walkers::{sources::Attribution, Map, MapMemory, Plugin, Tiles, TilesManager};
+use walkers::{Map, MapMemory, Plugin, Tiles, TilesManager};
 use {
     egui::{self, Context},
     re_entity_db::EntityProperties,
@@ -11,7 +10,6 @@ use {
     re_space_view::suggest_space_view_for_each_entity,
     re_types::blueprint::components::MapProvider,
     re_types::{SpaceViewClassIdentifier, View},
-    re_ui,
     re_viewer_context::{
         SpaceViewClass, SpaceViewClassLayoutPriority, SpaceViewClassRegistryError, SpaceViewId,
         SpaceViewSpawnHeuristics, SpaceViewState, SpaceViewStateExt as _,
@@ -21,15 +19,7 @@ use {
 };
 
 use crate::map_visualizer_system::{MapEntry, MapVisualizerSystem};
-
-// #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Default)]
-// pub enum MapProvider {
-//     #[default]
-//     OpenStreetMap,
-//     MapboxStreets,
-//     MapboxDark,
-//     MapboxSatellite,
-// }
+use crate::map_windows;
 
 // walkers plugin to visualize points on a Map
 pub struct PositionsOnMap {
@@ -248,44 +238,12 @@ impl SpaceViewClass for MapSpaceView {
                 }),
             );
 
-            // TODO(tfoldi): move to function along with acknowledge in a separate file
             let map_pos = map_widget.rect;
-            Window::new("Zoom")
-                .collapsible(false)
-                .resizable(false)
-                .title_bar(false)
-                .current_pos(egui::Pos2::new(map_pos.max.x - 80., map_pos.min.y + 10.))
-                .show(ui.ctx(), |ui| {
-                    ui.horizontal(|ui| {
-                        if ui.button(egui::RichText::new("➕").heading()).clicked() {
-                            let _ = map_memory.zoom_in();
-                        }
-
-                        if ui.button(egui::RichText::new("➖").heading()).clicked() {
-                            let _ = map_memory.zoom_out();
-                        }
-                    });
-                });
-            acknowledge(ui, &map_pos, tiles.attribution());
+            map_windows::zoom(ui, &map_pos, map_memory);
+            map_windows::acknowledge(ui, &map_pos, tiles.attribution());
         });
         Ok(())
     }
-}
-
-pub fn acknowledge(ui: &Ui, map_pos: &egui::Rect, attribution: Attribution) {
-    Window::new("Acknowledge")
-        .collapsible(false)
-        .resizable(false)
-        .title_bar(false)
-        .current_pos(egui::Pos2::new(map_pos.min.x + 10., map_pos.max.y - 40.))
-        .show(ui.ctx(), |ui| {
-            ui.horizontal(|ui| {
-                if let Some(logo) = attribution.logo_light {
-                    ui.add(egui::Image::new(logo).max_height(30.0).max_width(80.0));
-                }
-                ui.hyperlink_to(attribution.text, attribution.url);
-            });
-        });
 }
 
 fn get_tile_manager(provider: MapProvider, mapbox_access_token: &str, egui_ctx: &Context) -> Tiles {
